backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter~=23.3
djoser==2.1.0
psycopg2-binary==2.9.3
Pillow==9.0.0
PyYAML==6.0
webcolors==1.11.1
fpdf2==2.7.6

# pip install django-admin-list-filter-dropdown
